requirements.txt
langchain==1.2.10
langchain-core==1.2.13
langchain-community==0.4.1
langchain-openai==1.1.9
langchain-text-splitters
langgraph==1.0.8
openai
python-dotenv
faiss-cpu
tiktoken
pypdf
beautifulsoup4
requests
pydantic
